GOPRO: Set GoPro clock automatically
User or app needs to change /usr/bin/extSettings.conf "timeOffsetHrs" to be their local time offset from UTC in hours. The default is "None", and the clock will not update. The Solo has no other way to know what timezone it is in, so this must be set manually or with the app.  If set, for example to -5, the GoPro clock will automatically be set for the current UTC time minus 5 hours (eastern standard time).