GoPro: Use ArduCopter shutter if GoPro not connected
This will allow any other camera connected to the Solo to be triggered by the Solo's shutter controls if no GoPro is connected.  So, if no GoPro is connected, it will send the mavlink command for triggering the shutter in ArduCopter.  Whatever servo output have defined as a camera trigger will fire.