bugfixes for error/s that were occuring in 'vagrant up'.
this pins the git version used to 2.32.2 as it seems to work ok with the patches, and latest one from releases didn't. 
it also changes the url schema to http: from git: as per github guidlines change/s.
bitbake recipies need protocol=http, to tell bitbake that git
should use the http protocol